Proof of concept: finish and add parcel build
