+++ package.json
@@ -5,7 +5,8 @@
   "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "nodemon src/index.js"
+    "start": "node src/index.node.js && npx parcel src/index.html",
+    "node-start": "nodemon src/index.node.js"
   },
   "repository": {},
   "keywords": [],
@@ -16,10 +17,13 @@
   },
   "homepage": "https://github.com/tachyonizer/tachyonizer#readme",
   "dependencies": {
+    "@babel/preset-react": "^7.0.0",
     "express": "^4.16.4",
-    "nodemon": "^1.18.6"
+    "nodemon": "^1.18.6",
+    "parcel-bundler": "^1.10.3"
   },
   "devDependencies": {
+    "@babel/core": "^7.1.6",
     "@babel/generator": "^7.1.6",
     "@babel/parser": "^7.1.6",
     "@babel/traverse": "^7.1.6",
