+++ src/Title.css
@@ -1,0 +1,1 @@
+.tachyonizer {font-size: 32px;margin: 10px;border: 2px solid black;}
