+++ src/Title.js
@@ -1,0 +1,4 @@
+import "./Title.css";
+export default function Title() {
+  return <h1 className="tachyonizer">Тест</h1>;
+}
